Fixed the "backed up" icon for clip tiles which was only appearing while a clip was actively being uploaded to an FTP server.  I'm not sure when this broke, but I swear it was working when I wrote it initially.  Just like in the local console's clip list, there are now two icons: a cloud with an up-arrow for clips that are currently being backed-up, and a cloud with NO arrow for clips that are fully backed-up.  Since the clip list doesn't reliably (automatically) update the status of existing clip tiles, I'm not sure it makes a lot of sense to differentiate these two statuses like this, but it is what it is.
The clip properties panel can now display the flag for clips that are still recording, and the flag for alerts that were created while the camera was in a no-signal state.  These flags are not currently representable by icons on clip tiles, though if someone asks for either of them they could be added easily.  They do take up space in the settings panel if they are added.