Version 116: * To improve stability, the maximum keyframe interval in any UI3 streaming profile is reduced to 1000, from 99999.  This can be overridden locally by setting `settings.ui3_maxGOP = N`. * UI3's default VBR profiles now have a keyframe interval of 1000, down from 3000.
Tests have shown the HTML5 player in chrome to eventually freeze when a keyframe is not received for a long time, causing the orange clock icon to appear until the stream reset after being frozen for 60 seconds.  Testing also showed the player would freeze regardless of keyframe interval if the bit rate limit was disabled and the quality was set high.